Refactor: rename http_zip → http_archive, move audio gen, split detector find routes
Three independent reorgs:
- Rename vtsearch/datasets/importers/http_zip/ → http_archive/ (the API/CLI
  name was already http_archive; the directory name was a leftover)
- Move vtsearch/audio/generator.py → vtsearch/utils/audio_generator.py and
  delete the empty vtsearch/audio/ package (it was synthetic audio
  generation, not a media-type plugin)
- Extract /api/find, /api/find/check-labels, /api/find/progress from
  detectors_training.py into a new detectors_find.py blueprint;
  multi_find() was 348 lines and is its own subsystem (cross-dataset
  scoring), not training

Backwards compatibility broken: any external code importing
vtsearch.datasets.importers.http_zip or vtsearch.audio must update.
