+++ requirements-plugins.txt
@@ -6,7 +6,7 @@
 
 -r vtsearch/datasets/importers/combine_datasets/requirements.txt
 -r vtsearch/datasets/importers/folder/requirements.txt
--r vtsearch/datasets/importers/http_zip/requirements.txt
+-r vtsearch/datasets/importers/http_archive/requirements.txt
 -r vtsearch/datasets/importers/pickle/requirements.txt
 -r vtsearch/exporters/email_smtp/requirements.txt
 -r vtsearch/exporters/gui/requirements.txt
