+++ tests/helpers.py
@@ -20,7 +20,7 @@ def make_wav_bytes(frequency: float = 440.0, duration: float = 0.1) -> bytes:
     Supports variable ``frequency`` and ``duration`` — useful when tests
     need multiple distinct WAV files (e.g. different frequencies per media).
     """
-    from vtsearch.audio import generate_wav  # noqa: PLC0415
+    from vtsearch.utils.audio_generator import generate_wav  # noqa: PLC0415
 
     return generate_wav(frequency, duration)
 
@@ -29,7 +29,7 @@ def make_raw_wav_bytes() -> bytes:
     """Create a minimal valid WAV file (100 zero-samples) in memory.
 
     Lighter weight than :func:`make_wav_bytes` — does not depend on the
-    ``vtsearch.audio`` module.
+    ``vtsearch.utils.audio_generator`` module.
     """
     buf = io.BytesIO()
     with wave.open(buf, "wb") as wf:
