--- vtsearch/audio/__init__.py
@@ -1,7 +1,0 @@
-"""Audio generation utilities."""
-
-from vtsearch.audio.generator import generate_wav
-
-__all__ = [
-    "generate_wav",
-]
